requirements.txt
psycopg2~=2.8.6
flask~=1.1.2
matplotlib~=3.2.2
scikit-learn~=0.23.1
numpy~=1.18.5
scipy~=1.5.0
lxml~=4.5.2
opencv-python~=4.4.0.44
pandas~=1.0.5
pytesseract~=0.3.6
pillow~=7.2.0